REFACTOR : 컬렉션 Getter 삭제
