+++ README.md
@@ -3,18 +3,20 @@
 
 # 기능 명세
 * [x] 게임 플레이어들의 이름을 입력받는다.
-    * [ ] 입력 중 쉼표가 아닌 기호가 있을 경우의 예외 처리
-* [ ] 딜러와 플레이어들에게 각각 2장의 카드를 랜덤으로 나누어준다.
-    * [ ] 딜러의 카드 중 한 장만을 보여준다.
-    * [ ] 플레이어들의 카드들을 보여준다.
-* [ ] 플레이어들에게 카드를 더 받을 지 여부를 묻는다.
-    * [ ] 카드를 더 받고자 할 시, 랜덤으로 배분한다.
-    * [ ] 카드 배분이 끝나면 플레이어가 가지고 있는 카드를 보여준다.
-* [ ] 딜러의 카드의 합이 16이하인지 확인한다.
-    * [ ] 이하일 경우 1장의 카드를 추가로 받는다.
-* [ ] 딜러와 플레이어들 각각의 카드 합산을 계산한다.    
-    * [ ] 딜러와 플레이어들의 카드 현황과 합산 결과를 보여준다.
-* [ ] 최종 승패를 계산한다.
-    * [ ] 딜러는 각 플레이어와의 승패 결과를 나타낸다.
-    * [ ] 플레이어는 딜러와 결과 값을 비교하여 승패 결과를 나타낸다.
+    * [x] 입력 중 쉼표가 아닌 기호가 있을 경우의 예외 처리
+* [x] 카드 세팅
+* [x] 딜러와 플레이어들에게 각각 2장의 카드를 랜덤으로 나누어준다.
+    * [x] 딜러의 카드 중 한 장만을 보여준다.
+    * [x] 플레이어들의 카드들을 보여준다.
+* [x] 플레이어들에게 카드를 더 받을 지 여부를 묻는다.
+    * [x] 카드를 더 받고자 할 시, 랜덤으로 배분한다.
+    * [x] 카드 배분이 끝나면 플레이어가 가지고 있는 카드를 보여준다.
+* [x] 카드들의 포인트 합산을 계산한다.
+* [x] 딜러의 카드의 합이 16이하인지 확인한다.
+    * [x] 이하일 경우 1장의 카드를 추가로 받는다.
+* [x] 딜러와 플레이어들 각각의 카드 합산을 계산한다.    
+    * [x] 딜러와 플레이어들의 카드 현황과 합산 결과를 보여준다.
+* [x] 최종 승패를 계산한다.
+    * [x] 딜러는 각 플레이어와의 승패 결과를 나타낸다.
+    * [x] 플레이어는 딜러와 결과 값을 비교하여 승패 결과를 나타낸다.
